refactor: move departures layout
